+++ src/todos.js
@@ -1,5 +1,5 @@
-import displayTaks from './displayTsks';
 import { getTasksFromLocalStorage, addToLocalStorage } from './localStorage';
+import displayTasks from './displayTasks';
 
 export default class Todos {
   constructor() {
@@ -9,17 +9,15 @@ export default class Todos {
   renderTodos() {
     this.TodoList.sort((a, b) => a.index - b.index).forEach((todo) => {
       const checked = todo.completed ? 'checked' : '';
-      displayTaks(todo.index, todo.description, checked);
+      displayTasks(todo.index, todo.description, checked);
     });
   }
 
-  addTodoTask() {
+  addTodoTask(todo) {
     const id = this.TodoList.length;
-    const input = document.getElementById('input');
-    const todo = input.value;
 
     if (todo) {
-      displayTaks(id, todo, false);
+      displayTasks(id, todo, false);
 
       this.TodoList.push({
         index: id,
@@ -28,22 +26,7 @@ export default class Todos {
       });
 
       addToLocalStorage(this.TodoList);
-      input.value = '';
     }
-  }
-
-  addTodo(todo, id, status = false) {
-    const checked = status ? 'checked' : '';
-
-    displayTaks(id, todo, checked);
-
-    this.TodoList.push({
-      index: id,
-      description: todo,
-      completed: false,
-    });
-
-    addToLocalStorage(this.TodoList);
   }
 
   completeTodo(element) {
@@ -57,8 +40,8 @@ export default class Todos {
     addToLocalStorage(this.TodoList);
   }
 
-  removeTodo(element) {
-    const currentTodoId = parseInt(element.id, 10);
+  removeTodo(id) {
+    const currentTodoId = parseInt(id, 10);
 
     this.TodoList = this.TodoList.filter(
       (todo) => todo.index !== currentTodoId,
@@ -67,9 +50,10 @@ export default class Todos {
     for (let i = currentTodoId; i < this.TodoList.length; i += 1) {
       this.TodoList[i].index -= 1;
     }
-    element.parentNode.parentNode.replaceChildren('');
-    this.renderTodos();
+
     addToLocalStorage(this.TodoList);
+
+    this.renderTodos();
   }
 
   updateTodo(element) {
@@ -87,7 +71,7 @@ export default class Todos {
         parent.style.backgroundColor = 'transparent';
         element.classList = 'fa-solid fa-ellipsis-vertical';
         element.setAttribute('name', 'update');
-        this.TodoList[element.id].description = event.target.value;
+        this.TodoList[parseInt(element.id, 10)].description = event.target.value;
 
         addToLocalStorage(this.TodoList);
       }
